STATBLOCK ADDITION
Added a StatBlock class to Entity which stores the stats including level, health, attack, defense, and speed. This also
stores temporary values for battle purposes including defense and attack and speed buffs. This will be massively important
for button presses, particularly clicking the defense button.

This also will allow things to be taken out of the entity class once things are better refactored and optimized. One thing
that will also have to be refactored is calls to effective stats (base + modifiers) versus calls to base stats, if any
calls to base stats will be made.

More can be done later, but for now this will just be for defense, which is all that has been added (true values etc). First,
I will do refactoring, then merge this to the button functionality branch.