STATBLOCK AND ENEMY + PLAYER + SLIME REFACTORING
Using the new statblock architecture, refactored the entity class, enemy class, and player class, and
slime class, to all use it. Entity is now a much less cluttered class, and all stat calculations are
handled by the StatBlock object within the entities. Everything still works as intended, but it is now
going through the StatBlock class.

Will still need to refactor to include the effective stat retrieval in all stat calculations, but
for now this is okay. I may do that later or in the next commit.