DEFENSE BUTTON FIRST DRAFT (+ Entity refactoring to match)
This commit refactors the enemy and attackEntity method to use the effective defense, which is gotten from the temporary modifier. It also
adds the functionality of adding defense to the defense button, as well as resetting the defense modifier when the turn is over using
invokeLater, which hopefully is correct.

There is also a check in the entity guard method to check for the case of 0 defense, which will just add one to the temporary defense mod.

More refactoring must obviously be done, but it's in a good spot! With this, the next steps can be taken to add functionality to the other buttons.