STATBLOCK ADDITION (#3)
* STATBLOCK ADDITION

Added a StatBlock class to Entity which stores the stats including level, health, attack, defense, and speed. This also
stores temporary values for battle purposes including defense and attack and speed buffs. This will be massively important
for button presses, particularly clicking the defense button.

This also will allow things to be taken out of the entity class once things are better refactored and optimized. One thing
that will also have to be refactored is calls to effective stats (base + modifiers) versus calls to base stats, if any
calls to base stats will be made.

More can be done later, but for now this will just be for defense, which is all that has been added (true values etc). First,
I will do refactoring, then merge this to the button functionality branch.

* STATBLOCK AND ENEMY + PLAYER + SLIME REFACTORING

Using the new statblock architecture, refactored the entity class, enemy class, and player class, and
slime class, to all use it. Entity is now a much less cluttered class, and all stat calculations are
handled by the StatBlock object within the entities. Everything still works as intended, but it is now
going through the StatBlock class.

Will still need to refactor to include the effective stat retrieval in all stat calculations, but
for now this is okay. I may do that later or in the next commit.

* DEFENSE BUTTON FIRST DRAFT (+ Entity refactoring to match)

This commit refactors the enemy and attackEntity method to use the effective defense, which is gotten from the temporary modifier. It also
adds the functionality of adding defense to the defense button, as well as resetting the defense modifier when the turn is over using
invokeLater, which hopefully is correct.

There is also a check in the entity guard method to check for the case of 0 defense, which will just add one to the temporary defense mod.

More refactoring must obviously be done, but it's in a good spot! With this, the next steps can be taken to add functionality to the other buttons.