RUN BUTTON PREPARATION
Added in the entity StatBlock class get effective methods for all necessary stats which returns them plus the modifiers, which will be very
useful; we may not even need the standard get functions eventually.

STILL NEED TO REFACTOR ALL USAGES OF THE GET FUNCTIONS TO BE CORRECT. Otherwise, runFunction has been started but not yet added to the button.