RUN BUTTON FIRST DRAFT
This commit has the first draft of the run button. It involves printing successful or failed run, then pausing using the new thing I just learned of a GUI
timer! That will definitely be something I'll have to look more into in the future.

After this timer is elapsed, the system exits.

There are some bugs that must be fixed: the enemy still attacks player after they have run, which should not happen, but should be solved by adding some
checks about being in battle (finally I get to actually use that in battle check!). The second bug is that the player can still use inputs, which should
have been solved by setting player turn to false, but I'm not sure why that's not working; it might be fixed again by setting inBattle to false, but I
want to find out why it's not working with the playerTurn boolean.

But, the run button generally does work; just needs some fixing! You can also fail or succeed the run, although a test will need to be implemented to make
sure that is working as expected.