ONE TURN, NEW ARCHITECTURE
This may seem like a downgrade, but it is all in the name of better functionality. The turn system was an absolute abhorrent mess. So, instead, I am restructuring it to make
room and better flow of button management. This will help fix bugs like enemies attacking when running is done, etc. It will also just make the battle better and easier to
refactor and make better in the future.

Once this refactoring is done, I will fix the bugs with running and then I will actually move on to adding the item button functionality.