Exiting Architecture
Some exit architecture was added this time around. I'm not sure why I went on and did this, I kind of just got sucked into the wormhole I
created. That's not great. But now I'm going to move on to making this a multi-turn situation. Oh yeah, I wanted to do this to make sure
inBattle was working and how to get it to work, since that determines when the turn is done.

Anyway, I forgot to mention last commit this is currently a one-turn situation, hence the downgrade. But by adding the runBattleTurn at the end
of every player input and enemy move, I should be able to then have moves go on until the player or slime is dead.