REFACTORED BATTLE + TEST FOR GOING FIRST
This commit involves a very much refactored battle controller, which in turn makes the battle MUCH easier to understand and deal with. What essentially happens now is that
the battle scene stores who the first goer is, and the battle controller changes that based on determineFirst. If the player goes first, it runs enemy turn, but if not, it
runs the determineFirst method again. Vice versa with the enemy. This seems to be working! Hallelujah! It only took deleting and rewriting the entire battle controller class
for this to be implemented well.

I also created a tester for this determineFirst method, that runs as many times as you'd like and gives you genuine success and failure messages depending on if it matches
with the expected values. And... IT DOES!!! I even have nice green text for when a test is successful. So that determineFirst is working to return the correct Entity based
on if the speed is lower, the same, or higher than the player's.

There are two steps left; the first is to get a genuine battle end function going, which involves making the inBattle boolean actually do something in battle controller. The
next will be making the run button interact with that in such a way that it prevents the second bug; the enemy attacking the player after they select running from the battle.
Once the first part is done (having inBattle do something), I will commit the final version to this branch and then the next commit will be back on the BUTTON-FUNCTIONALITY
branch.

Anyway these were big changes, and honestly a lot more involved than I expected, but hopefully this all goes well.