REFACTORING BATTLE SCENE AND ADJACENT -- final before merge
This commit was ALL about refactoring. I added local variables to methods, shortened them, made them over all more readable. In fact, these classes are actually looking fairly
decent right now. So, they're in a spot where I can go back to actually adding button functionality.

I also bug fixed the run methods to end the battle properly and prevent anything from happening once the run is successful. I did this by putting the endBattle method in the
handleRun stuff, and putting it before the end Player Turn, so the isInBattle bool is false when it gets to ending the player turn, which prevents it from running the enemy
turn. This comes with the caveat that the endPlayerTurn method must be rather simple, or must include a second check for if battle is over; which isn't too bad, but still might
interact all funny. So I'll have to keep that in mind.

Finally, I can move on to doing Item stuff for the item button, which will be interesting... definitely only the player will have an inventory for now. And an empty one. Sorry
'bout it.