ONE TURN, NEW ARCHITECTURE (#4)
* ONE TURN, NEW ARCHITECTURE

This may seem like a downgrade, but it is all in the name of better functionality. The turn system was an absolute abhorrent mess. So, instead, I am restructuring it to make
room and better flow of button management. This will help fix bugs like enemies attacking when running is done, etc. It will also just make the battle better and easier to
refactor and make better in the future.

Once this refactoring is done, I will fix the bugs with running and then I will actually move on to adding the item button functionality.

* Exiting Architecture

Some exit architecture was added this time around. I'm not sure why I went on and did this, I kind of just got sucked into the wormhole I
created. That's not great. But now I'm going to move on to making this a multi-turn situation. Oh yeah, I wanted to do this to make sure
inBattle was working and how to get it to work, since that determines when the turn is done.

Anyway, I forgot to mention last commit this is currently a one-turn situation, hence the downgrade. But by adding the runBattleTurn at the end
of every player input and enemy move, I should be able to then have moves go on until the player or slime is dead.

* REFACTORED BATTLE + TEST FOR GOING FIRST

This commit involves a very much refactored battle controller, which in turn makes the battle MUCH easier to understand and deal with. What essentially happens now is that
the battle scene stores who the first goer is, and the battle controller changes that based on determineFirst. If the player goes first, it runs enemy turn, but if not, it
runs the determineFirst method again. Vice versa with the enemy. This seems to be working! Hallelujah! It only took deleting and rewriting the entire battle controller class
for this to be implemented well.

I also created a tester for this determineFirst method, that runs as many times as you'd like and gives you genuine success and failure messages depending on if it matches
with the expected values. And... IT DOES!!! I even have nice green text for when a test is successful. So that determineFirst is working to return the correct Entity based
on if the speed is lower, the same, or higher than the player's.

There are two steps left; the first is to get a genuine battle end function going, which involves making the inBattle boolean actually do something in battle controller. The
next will be making the run button interact with that in such a way that it prevents the second bug; the enemy attacking the player after they select running from the battle.
Once the first part is done (having inBattle do something), I will commit the final version to this branch and then the next commit will be back on the BUTTON-FUNCTIONALITY
branch.

Anyway these were big changes, and honestly a lot more involved than I expected, but hopefully this all goes well.

* REFACTORING BATTLE SCENE AND ADJACENT -- final before merge

This commit was ALL about refactoring. I added local variables to methods, shortened them, made them over all more readable. In fact, these classes are actually looking fairly
decent right now. So, they're in a spot where I can go back to actually adding button functionality.

I also bug fixed the run methods to end the battle properly and prevent anything from happening once the run is successful. I did this by putting the endBattle method in the
handleRun stuff, and putting it before the end Player Turn, so the isInBattle bool is false when it gets to ending the player turn, which prevents it from running the enemy
turn. This comes with the caveat that the endPlayerTurn method must be rather simple, or must include a second check for if battle is over; which isn't too bad, but still might
interact all funny. So I'll have to keep that in mind.

Finally, I can move on to doing Item stuff for the item button, which will be interesting... definitely only the player will have an inventory for now. And an empty one. Sorry
'bout it.