NEW SYSTEMS! INVENTORY AND ITEM BEGINNINGS
This commit adds the beginnings of the inventory and item system. It just adds the item class, which will be populated with methods just for the inventory class to use, and the basic and helper
methods for the inventory class and constructor.

Really, I just need the inventory to exist and be possible to be empty for the button functionality, so that's all I'm really going to be doing here. I will add some basic infrastructure in this
branch, and end it with adding these things to the player class.