INVENTORY ADDITION AND BEGINNING ITEM BUTTON FUNCTIONALITY
In this commit, the inventory object was added to the player object with basic methods and helper methods to add future functionality. Add, remove, remove at index, etc. have all
been added. More should be added, perhaps a sort ability, etc., but future updates can be made in a dedicated branch.

This commit also adds basic functionality to the items button, though full functionality is yet to come. All it does is check if the inventory is empty. If it is, then it logs a
message to the battle panel but importantly does NOT end the players turn; so the enemy doesn't attack, and the player can do something else with their turn. If it is not, then
nothing happens. Though this isn't an issue for the tutorial, I do want to add genuine functionality, which will involve adding more to the items and inventory systems. Then, the
final commit of this branch can be a genuine item functionality thing. And maybe a test, too, if I'm feeling crazy.