HEALABLE INFRASTRUCTURE ADDED + GIVEN TO PLAYER
This commit adds healers as a class extending item with the only parameter being a (temporary) enum that will be a list of healing items. It takes in that healing item and uses
its data as its fields' values. In the tester, the player is also given the only item that exists––a small health potion––to test its healing effects. A use method must still be added
to the items, and I'm going to commit again after that is added as a smaller more atomic commit.

Once the use method is added and working properly, I'll add it to the item button, and finally the button functionality branch will be complete! Then I'm going to make a new release.
Then I'm going to spend some time refactoring EVERYTHING in a refactoring branch. Then this will be a genuine, actually fully functional MVP! After that, I can start adding functions.
EXP gain, a continuation of battle, a window outside of battle + battle exits, and essentially build an entire battle engine. I can build NPCs, I can go wherever the hell I'd like!