HEALABLE FUNCTIONALITY ADDED, ITEM BUTTON FUNCTIONAL, MINOR STAT BLOCK REFACTORING
This commit first adds healable functionality, which actually heals the player a set amount using the enum values (but by creating a unique object from that enum). To make this work,
I added a getUsed method to items. It returns int, but I'll have to see how to refactor that when the time comes for more item architecture. The item button is also now functional,
although it is missing a print statement for when items are actually used. The Stat Block class was also minorly refactored with clamps on the damage and heal methods, although this
hasn't been tested or debugged to make sure it works properly. Also added a takeHeal method to the stat block.

A few things to note:
1. There are a lot of temporary fixes and hard coded values as a band-aid to the deeper need to re-architect the item functionality, but that will come later.
2. There is a glitch where the player remains at 0 health and never loses (but luckily doesn't go below 0 health!). This should be an easy fix, though; just add a check for it in the
enemy attack method. I just overlooked that.

Next commit goals: BugFix for #2 and print statements for using items. Then this branch can be finished.