NEW SYSTEMS! INVENTORY AND ITEM BEGINNINGS (#5)
* NEW SYSTEMS! INVENTORY AND ITEM BEGINNINGS

This commit adds the beginnings of the inventory and item system. It just adds the item class, which will be populated with methods just for the inventory class to use, and the basic and helper
methods for the inventory class and constructor.

Really, I just need the inventory to exist and be possible to be empty for the button functionality, so that's all I'm really going to be doing here. I will add some basic infrastructure in this
branch, and end it with adding these things to the player class.

* INVENTORY ADDITION AND BEGINNING ITEM BUTTON FUNCTIONALITY

In this commit, the inventory object was added to the player object with basic methods and helper methods to add future functionality. Add, remove, remove at index, etc. have all
been added. More should be added, perhaps a sort ability, etc., but future updates can be made in a dedicated branch.

This commit also adds basic functionality to the items button, though full functionality is yet to come. All it does is check if the inventory is empty. If it is, then it logs a
message to the battle panel but importantly does NOT end the players turn; so the enemy doesn't attack, and the player can do something else with their turn. If it is not, then
nothing happens. Though this isn't an issue for the tutorial, I do want to add genuine functionality, which will involve adding more to the items and inventory systems. Then, the
final commit of this branch can be a genuine item functionality thing. And maybe a test, too, if I'm feeling crazy.

* HEALABLE INFRASTRUCTURE ADDED + GIVEN TO PLAYER

This commit adds healers as a class extending item with the only parameter being a (temporary) enum that will be a list of healing items. It takes in that healing item and uses
its data as its fields' values. In the tester, the player is also given the only item that exists––a small health potion––to test its healing effects. A use method must still be added
to the items, and I'm going to commit again after that is added as a smaller more atomic commit.

Once the use method is added and working properly, I'll add it to the item button, and finally the button functionality branch will be complete! Then I'm going to make a new release.
Then I'm going to spend some time refactoring EVERYTHING in a refactoring branch. Then this will be a genuine, actually fully functional MVP! After that, I can start adding functions.
EXP gain, a continuation of battle, a window outside of battle + battle exits, and essentially build an entire battle engine. I can build NPCs, I can go wherever the hell I'd like!

* HEALABLE FUNCTIONALITY ADDED, ITEM BUTTON FUNCTIONAL, MINOR STAT BLOCK REFACTORING

This commit first adds healable functionality, which actually heals the player a set amount using the enum values (but by creating a unique object from that enum). To make this work,
I added a getUsed method to items. It returns int, but I'll have to see how to refactor that when the time comes for more item architecture. The item button is also now functional,
although it is missing a print statement for when items are actually used. The Stat Block class was also minorly refactored with clamps on the damage and heal methods, although this
hasn't been tested or debugged to make sure it works properly. Also added a takeHeal method to the stat block.

A few things to note:
1. There are a lot of temporary fixes and hard coded values as a band-aid to the deeper need to re-architect the item functionality, but that will come later.
2. There is a glitch where the player remains at 0 health and never loses (but luckily doesn't go below 0 health!). This should be an easy fix, though; just add a check for it in the
enemy attack method. I just overlooked that.

Next commit goals: BugFix for #2 and print statements for using items. Then this branch can be finished.