ITEM BUTTON FUNCTIONAL
Item button is now functional with a minor change of moving useItem to player class since players are the only ones with inventories. Now, the
item is properly used when pressing the Item button, which seems like full functionality to me.

There still need to be bug fixes and hard-coded value fixes, but after those this will be good to go back into main.