BUGFIX INFINITE PLAYER 0 HEALTH + REFACTORING
This commit fixes the infinite player 0 health glitch as well as does some minor refactoring to methods by adding local variables. More refactoring must be
done of course, but this is a small atomic commit that slightly includes this stuff.

With that, this branch is done and can be merged into main! :-)