GAME PANEL + KINDA MESSED UP (#6)
* GAME PANEL + KINDA MESSED UP

This commit starts adding the infrastructure around and outside the battle scene by creating the game panel that will serve as the "shell" from which the
battle panel is created. Essentially, the game panel is an abstract class of all gui scenes that will take place (here, scenes will be referred to as
panels). The variables and fields as well as constructor are created, but the actual logic and situation needs to be added as well as how each subclass
will inherit from this abstract super.

Finally, I messed up a bit. I tried adding maven, then realized I really didn't want to have to deal with it, so now there are these .xml files in my
commit that I'm not sure if they're needed. I tried deleting them but everything broke, so I'm just going to commit them and hope things aren't broken...

* TEXT LOG SEPARATED OUT

This commit creates a new class for the Text Log, so rather than only have it as a generic JTextArea in the GamePanel, it is an actual wrapper class of its own that can
contain all the necessary log and print methods. In here will be generic print methods that will be used in GamePanel and BattlePanel, hopefully with some better logic
to make them more readable, refactored, and generic. However, this commit just sets up the basic constructor and methods, and it is not yet implemented.

* GAME PANEL FUNCTIONAL, BATTLE PANEL EXTENDS + TEXT LOG CLASS

This commit refactors a lot of GUI stuff, creating a separate class for the text log and moving many methods there. This also creates the generic game panel class, which will never be instantiated probably, but
instead serves as a super clas for the battle panel and then the story panel so that variables do not need to be repeated. Many methods were moved around, but the functionality appears to be the same throughout.
This will allow story panel and battle panel to share the same textLog, since they don't instantiate their own but rather use the one from GamePanel.

However, I have to check on that; if they extend GamePanel, and GamePanel creates a new textLog, wouldn't that still mean a new one is created every time the subclasses are created? I'll have to check on that.

* BUTTON REORGANIZATION

This commit reorganizes how the buttons work. Rather than Action Listeners being created and set up in the BattleController, now there is an enum that takes in a label for the button and calls upon the
BattleController to do some action. When the battle panel (A new class) is created, it goes through the BattleButton enum and creates new buttons with those titles and actions, adding them to itself.

This does have one issue: the tighter coupling from the fact we need to pass the controller into the GUI. This might make things tricky, and maybe in future I will refactor this to make it less evil, but for now
it will work and saves me the hassle of having to define each button for each type of GUI scene we have. Now, I'm going to move on and do that for the other game gui things.

* MAIN MENU FIRST STEPS

I meant to commit MUCH earlier and much more atomically. I'm still getting used to that. This commit has a lot of the main menu infrastructure created including its buttons and its button panel. There's
still the issue of tight coupling which is occurring between the GUI and controller, but I'll have to refactor that later. For now, I just want to get this up and running and then refactor including
abstracting out the button panel stuff to the general button panel class.

I WILL DO ATOMIC COMMITS FROM HERE ON OUT!! I PROMISE!!!

* MAIN MENU WORKING

Main menu is now working and buttons work properly, outputting to the console when pressed. Next will be creating the storyGUI that New Game will access.

* FOLDER REORGANIZATION

Moved everything into model, view, and controller packages to make things a little easier and more MVC-oriented.

* ADDITIONAL SUBCLASS - GAMEPLAY GUI

This commit adds a subclass to prevent logical weirdness by having main menu GUI and the actual gameplay GUI at the same level. So now the game play gui can hold the text log, whereas the main menu
can use other things that aren't text logs or to do with gameplay, reducing its bloat. All these files were refactored to move methods, etc, and it works.

* FOLDER REORGANIZATION

Did some more folder renaming, refactoring, and organization to make the code cleaner in the file viewer--if this really does that. The idea is that the scenes themselves are
model, whereas the guis are the view parts of it. Then there's a scene controller that manages them.

* FOLDER REORG + NEW GUI

Moved some stuff again, and then created the story button panel and story gui for the actual out of battle gameplay. The only buttons will be yes or no (for now), and I want to have it where the story panel
shows "listening..." when those buttons are not in use. But that will all go in these new classes and panels! I also may rename the ButtonPanel class to ActionPanel, since it won't just have buttons but will
be the main way of interacting with the game (perhaps providing movement in a future re-run of this gui engine...)

* BUG FIX + ADDED CLASS

Bug fix due to never actually adding a border layout to my gui panel, meaning the parts weren't being added to the layout correctly and the action panel would show up above the text log. That's now fixed.
The other thing I did was an add enum for GameChoice, which works for only yes or no questions. Something to be expanded in the future, but for now it's okay.

Another thing that this commit is demonstrating to me is that this has vastly escaped the original repo's original purpose of a battlescene MVP. This is turning much more into a Text-Based-Game Engine MVP. SO,
I'm going to try and move this branch over to a new repository just for that, and then use the battlescene mvp to actually flesh out the battle system more. What I'll probably do is de-bloat the version that
gets copied over.

I can either do this through intellij, or be really caveman about it and just copy and paste the files. Not sure which I'll do, but this will likely be the last commit on this branch.