PANEL MANAGEMENT - 2: Battle Display Begins
This commit begins detaching battle display from the old infrastructure and adding it to the new inheritance tree. This is definitely going to involve a decent amount of work, but it will definitely be worth it.

And then when I learn LibGDX it'll probably all not be worth it. But hey, I doubt it's a BAD thing to learn Java swing, right...? Right!??!