PANEL MANAGEMENT 4: SFSF - some formatting, some functionality
This commit creates the stat displayers for the two entities, and though they are placeholders right now, I think I will add specific ones for enemies and players when I get around to adding that knowledge
mechanic.

The important part, however, is that the panel is vaguely formatted as I want it AND the functionality is there. I just need to add the actual stat blocks for the entities in the stat displayers and once that's
done (by creating a properly formatted toString of entity stats, though I'll have to see how that interacts with that future knowledge mechanic I mentioned earlier). Now I can start moving to other things as well
like adding wrapper classes with formatting and functionality.