PANEL MANAGEMENT 5: Final battle panel
This commit finalizes the battle panel and its general format, with the enemy's and player's stats displayed, the text log in the center, and the button panel at the bottom. So far nothing does anything, but that
will be everything the next branch is for. I'm simultaneously doing a LOT of code reorganization to make things cleaner and more understandable and now with a better understanding of MVC principles I'm going to
be able to do much, much more.