PANEL MANAGEMENT (#12)
* PANEL MANAGEMENT

This first commit creates better panel infrastructure with a generic abstract content panel class that then the display and interact panels will extend. Then, individual scenes will have a display panel and a
button panel that extend those classes. Neat, huh?

Anyway, these aren't implemented yet but they will be in the next commit!

* PANEL MANAGEMENT - 2: Battle Display Begins

This commit begins detaching battle display from the old infrastructure and adding it to the new inheritance tree. This is definitely going to involve a decent amount of work, but it will definitely be worth it.

And then when I learn LibGDX it'll probably all not be worth it. But hey, I doubt it's a BAD thing to learn Java swing, right...? Right!??!

* PANEL MANAGEMENT 3: It... it works? The panel works?!

This commit adds the battle display panel without any bits and bobs, so just the text area taking up the entire border layout which I do want to change and I think will change with the addition of the side
text areas. It also adds the button panel at the bottom with the buttons successfully centered in a flow layout.

I'm learning... I'M LEARNING!!!! I'M BEGINNING TO UNDERSTAND SWING AND I'M ACTUALLY MAKING FUNCTIONAL, USABLE GUI!!!!!!!! MY OWN VERSION OF QUEST, HERE I COOOOOOOOOOME!!!!!!!!!!

* PANEL MANAGEMENT 4: SFSF - some formatting, some functionality

This commit creates the stat displayers for the two entities, and though they are placeholders right now, I think I will add specific ones for enemies and players when I get around to adding that knowledge
mechanic.

The important part, however, is that the panel is vaguely formatted as I want it AND the functionality is there. I just need to add the actual stat blocks for the entities in the stat displayers and once that's
done (by creating a properly formatted toString of entity stats, though I'll have to see how that interacts with that future knowledge mechanic I mentioned earlier). Now I can start moving to other things as well
like adding wrapper classes with formatting and functionality.

* PANEL MANAGEMENT 5: Final battle panel

This commit finalizes the battle panel and its general format, with the enemy's and player's stats displayed, the text log in the center, and the button panel at the bottom. So far nothing does anything, but that
will be everything the next branch is for. I'm simultaneously doing a LOT of code reorganization to make things cleaner and more understandable and now with a better understanding of MVC principles I'm going to
be able to do much, much more.