MODEL REFACTORING 1: Stat Block as Hash Map
This commit refactors the Stat Block from merely a collection of integers into a fully fledged Hash Map with everything in it. Things print properly, and I'm no longer printing the temporary stats. Instead, they
will just be added whenever they are created.

But this is the first step to making the entity logic much, much cleaner.

One thing I may do far in the future is make the stat block implement an interface for calculating stats. But for now it's all good. Next, I'm going to move on to fixing up the entity code itself outside of the
stats.