MODEL REFACTORING 4: Entity methods
Entity has now been populated with methods that use stat block methods to do certain things. This *should* hopefully be all the methods that entity needs, at least in regard to its stats and what it does in
battle.

There are also the placeholders for future expansion, including for using items as well as changing the state the entity is in (alive, dead, etc.).