MODEL REFACTORING 4: Better toStrings + LifeState
This commit adds better to String logic for the entity class and the stat block to more easily print the to strings. It is super useful and so much better to use String.format(). Eventually I'll move these
toStrings to logs and will have specific displayer string methods, but otherwise its good stuff.

This commit also adds a life state enum to the entity that will tell whether they are ALIVE or DEAD. This will be useful for enemy logic and stuff and is also displayed in the stat displayer.