MODEL REFACTORING 1: Stat Block as Hash Map (#14)
* MODEL REFACTORING 1: Stat Block as Hash Map

This commit refactors the Stat Block from merely a collection of integers into a fully fledged Hash Map with everything in it. Things print properly, and I'm no longer printing the temporary stats. Instead, they
will just be added whenever they are created.

But this is the first step to making the entity logic much, much cleaner.

One thing I may do far in the future is make the stat block implement an interface for calculating stats. But for now it's all good. Next, I'm going to move on to fixing up the entity code itself outside of the
stats.

* MODEL REFACTORING 2: Stat methods

This commit adds stats methods to the Stat block class regarding leveling up and down, taking damage and healing. These are the basic stat alterations that will be done for now, though more may be added in
future.

* MODEL REFACTORING 3: Stat Block and Entity calculation methods

This commit adds calculation methods for both the stat block and entity, including things like guarding and running. It adds comparison methods and increase methods for temporary stats. Good stuff. It also adds
better commenting which I'm going to try and start doing more using javadoc commenting now that I know what it means.

* MODEL REFACTORING 4: Entity methods

Entity has now been populated with methods that use stat block methods to do certain things. This *should* hopefully be all the methods that entity needs, at least in regard to its stats and what it does in
battle.

There are also the placeholders for future expansion, including for using items as well as changing the state the entity is in (alive, dead, etc.).

* MODEL REFACTORING 4: Better toStrings + LifeState

This commit adds better to String logic for the entity class and the stat block to more easily print the to strings. It is super useful and so much better to use String.format(). Eventually I'll move these
toStrings to logs and will have specific displayer string methods, but otherwise its good stuff.

This commit also adds a life state enum to the entity that will tell whether they are ALIVE or DEAD. This will be useful for enemy logic and stuff and is also displayed in the stat displayer.

* MODEL REFACTORING 5: Enemy and Player functions

This commit adds the enemy and player functions regarding their specific situations including leveling up and calculating EXP amounts. It adds how the player will level and how their stats will change on level up
which right now involves random numbers being added to their stats and just a linear function of increasing the experience required for levelling up by 50 times the level. Eventually I will add more complex stuff
but for the simple scope this is fine.