BATTLE STATE REFACTORING
This commit is the first in this branch that is refactoring how the battle is done. This creates a battle state class which will act as a model portion of the battle. It will return things like the entity that
goes first, the damage that entities do to each other, and some sort of internal logic about the battle scene itself as a thing. It might deal with attacking, but I'm not entirely sure. That might be in the
controller since it will deal with buttons. But there will probably be some level of attacking that is delegated to the battle scene class.

This is the first commit in the next phase of development; refactoring and making better the battle scene. Then, I can move on to the actual battle controller.