BATTLE STATE REFACTORING 2: Redone Battle State + Battle End Codes
This commit essentially redoes the battle state model class. It now encodes a player and enemy (which can be expanded in future with var args?), a boolean for the player turn to handle when the battle controller
tells the gui to enable buttons, and an end code so the controller can do things after the battle has ended depending on what happens to end the battle.

It also now has an attack entity method, which the controller can call on when necessary.

This commit also adds some necessary things to the entity class due to what has been added to the battle scene. Most notably, it adds a battle choice enum to the entity constructor, which allows entities to use
certain battle choices applied to them during the battle, which can then determine their actions in the controller.