BATTLE STATE REFACTORING (#15)
* BATTLE STATE REFACTORING

This commit is the first in this branch that is refactoring how the battle is done. This creates a battle state class which will act as a model portion of the battle. It will return things like the entity that
goes first, the damage that entities do to each other, and some sort of internal logic about the battle scene itself as a thing. It might deal with attacking, but I'm not entirely sure. That might be in the
controller since it will deal with buttons. But there will probably be some level of attacking that is delegated to the battle scene class.

This is the first commit in the next phase of development; refactoring and making better the battle scene. Then, I can move on to the actual battle controller.

* BATTLE STATE REFACTORING 2: Redone Battle State + Battle End Codes

This commit essentially redoes the battle state model class. It now encodes a player and enemy (which can be expanded in future with var args?), a boolean for the player turn to handle when the battle controller
tells the gui to enable buttons, and an end code so the controller can do things after the battle has ended depending on what happens to end the battle.

It also now has an attack entity method, which the controller can call on when necessary.

This commit also adds some necessary things to the entity class due to what has been added to the battle scene. Most notably, it adds a battle choice enum to the entity constructor, which allows entities to use
certain battle choices applied to them during the battle, which can then determine their actions in the controller.

* BATTLE STATE REFACTORING 3: Just a little change

Just changed something in the controller, now I'm going to finish this branch and start working on the controller.